Create needed TS bucket as part of test setup.
Moved test set into a function to defer running tests until after setup is run.
Previously, the tests were executed at the definition site, so they were run
before the setup step. This now makes the put that was in the setup fail, but
removing that doesn't cause any of the other tests to fail.

Also added more context to the is_testable_path so more helpful logging can be
shown.