Add integration tests for pb-messages endpoints
To make the tests idempotent, a check is needed before the tests are run to see
if the ProtoBuf file has already been uploaded. The endpoint returns a 200 if
the file is new, and a 204 if it's already in place. By adding this check, we
can expect the correct code.